Move files to common-compose-ui
