+++ android/src/main/java/dev/sergiobelda/todometer/ui/taskdetail/TaskDetailScreen.kt
@@ -61,10 +61,10 @@ import androidx.compose.ui.unit.dp
 import dev.sergiobelda.todometer.R
 import dev.sergiobelda.todometer.common.compose.ui.designsystem.components.HorizontalDivider
 import dev.sergiobelda.todometer.common.compose.ui.designsystem.components.ToDometerCheckbox
-import dev.sergiobelda.todometer.common.compose.ui.designsystem.mapper.composeColorOf
-import dev.sergiobelda.todometer.common.compose.ui.designsystem.task.TaskDueDateChip
-import dev.sergiobelda.todometer.common.compose.ui.designsystem.taskchecklistitem.AddChecklistItemField
 import dev.sergiobelda.todometer.common.compose.ui.designsystem.theme.ToDometerTheme
+import dev.sergiobelda.todometer.common.compose.ui.mapper.composeColorOf
+import dev.sergiobelda.todometer.common.compose.ui.task.TaskDueDateChip
+import dev.sergiobelda.todometer.common.compose.ui.taskchecklistitem.AddChecklistItemField
 import dev.sergiobelda.todometer.common.domain.model.Task
 import dev.sergiobelda.todometer.common.domain.model.TaskChecklistItem
 import dev.sergiobelda.todometer.common.domain.model.TaskChecklistItemState
