+++ common/compose-ui/src/commonMain/kotlin/dev/sergiobelda/todometer/common/compose/ui/mapper/ColorMapper.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.sergiobelda.todometer.common.compose.ui.designsystem.mapper
+package dev.sergiobelda.todometer.common.compose.ui.mapper
 
 import androidx.compose.runtime.Composable
 import dev.sergiobelda.todometer.common.compose.ui.designsystem.theme.ToDometerColors
