+++ common/compose-ui/src/commonMain/kotlin/dev/sergiobelda/todometer/common/compose/ui/tasklist/TaskListItem.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.sergiobelda.todometer.common.compose.ui.designsystem.tasklist
+package dev.sergiobelda.todometer.common.compose.ui.tasklist
 
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
