+++ desktop/src/jvmMain/kotlin/ui/home/HomeScreen.kt
@@ -57,9 +57,9 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import dev.sergiobelda.todometer.common.compose.ui.designsystem.components.VerticalDivider
-import dev.sergiobelda.todometer.common.compose.ui.designsystem.task.TaskItem
-import dev.sergiobelda.todometer.common.compose.ui.designsystem.tasklist.TaskListItem
-import dev.sergiobelda.todometer.common.compose.ui.designsystem.tasklist.TaskListProgress
+import dev.sergiobelda.todometer.common.compose.ui.task.TaskItem
+import dev.sergiobelda.todometer.common.compose.ui.tasklist.TaskListItem
+import dev.sergiobelda.todometer.common.compose.ui.tasklist.TaskListProgress
 import dev.sergiobelda.todometer.common.domain.doIfSuccess
 import dev.sergiobelda.todometer.common.domain.model.Tag
 import dev.sergiobelda.todometer.common.domain.model.TaskItem
